Added convienience static methods on XMLStreamBuffer to avoid the direct use of MutableXMLStreamBuffer for common cases. Optimized the space taken by the stream buffer infoset represenation. Using a byte[] for the structure (as originally designed). Lengths of character arrays are encoded as 8 bit or 16 bit non-negative integers in byte[]. On average this reduces the size of the structure array by a factor of 4 for very minimal overhead in creation and processing cost.
svn path=/trunk/; revision=110

+++ streambuffer/src/com/sun/xml/stream/buffer/MutableXMLStreamBuffer.java
@@ -83,7 +83,7 @@ public class MutableXMLStreamBuffer extends XMLStreamBuffer {
      * If the <code>size</code> argument is less than <code>0</code>.
      */
     public MutableXMLStreamBuffer(int size) {
-        _structure = new FragmentedArray(new int[size]);
+        _structure = new FragmentedArray(new byte[size]);
         _structureStrings = new FragmentedArray(new String[size]);
         _contentCharactersBuffer = new FragmentedArray(new char[4096]);
         _contentObjects = new FragmentedArray(new Object[size]);
